Implemented continuous ant colony optimization
Class ContinuousACO implements an ant system for solving continuous
optimization problems in a n-dimensional space. This class extends
the new EvolutionaryAlgorithm class which abstracts the common
parameters of an evolutionary algorithm for solving optimization
problems, such the objective function, domain, population initialization
etc.

A class RandomSelector was created to probabilistic select and item
from a list of items, useful in some EA's, such the selection of an
dimension interval in the ACO.