Now BasicBinaryGA extends EvolutionaryAlgorithm and uses RandomSelector
BasicBinaryGA now extends EvolutionaryAlgorithm so it does not
reimplement some methods, such population initialization, population
sorting or the fitness function. Some methods were overwritten to
work with the GA, such random_individual which generates a candidate
solution formed with genotypes values, and the evaluate method which
needs to map the genotypes into fenotypes before evaluating the
individual.

The "sample" method was added to RandomSelector which returns a list
of n unique items from the population. The "pick" method was renamed
to "choose" for clarity. The RandomSelector is used by BasicBinaryGA
to select two parents to mate in the crossover operation, thus the
"sample" method.