Refactor to EvolutionaryAlgorithm interface
Defined public and non-public members. Now its posible to specify
the lower and upper bounds for each dimension.

The evolve method does a single iteration; the solve methods calls
evolve the necessary times, this allows to sort and save the
population on each generation.

The _get_generation method returns a new Generation instance, this
method is called after each generation and the result is appended
to a generations array. This data is primary intended for plotting.